Define generic bridge pallets call structs
