+++ relays/bin-substrate/src/chains/wococo_parachains_to_bridge_hub_rococo.rs
@@ -51,7 +51,7 @@ impl SubmitParachainHeadsCallBuilder<BridgeHubWococoToBridgeHubRococo>
 		parachain_heads_proof: ParaHeadsProof,
 	) -> CallOf<relay_bridge_hub_rococo_client::BridgeHubRococo> {
 		relay_bridge_hub_rococo_client::runtime::Call::BridgeWococoParachain(
-			relay_bridge_hub_rococo_client::runtime::BridgeParachainCall::submit_parachain_heads(
+			bp_parachains::BridgeParachainCall::submit_parachain_heads(
 				(at_relay_block.0, at_relay_block.1),
 				parachains,
 				parachain_heads_proof,
