+++ relays/client-bridge-hub-rococo/Cargo.toml
@@ -16,8 +16,10 @@ bp-bridge-hub-rococo = { path = "../../primitives/chain-bridge-hub-rococo" }
 bp-bridge-hub-wococo = { path = "../../primitives/chain-bridge-hub-wococo" }
 bp-header-chain = { path = "../../primitives/header-chain" }
 bp-messages = { path = "../../primitives/messages" }
-bp-polkadot-core = { path = "../../primitives/polkadot-core" }
+bp-parachains = { path = "../../primitives/parachains" }
+bp-rococo = { path = "../../primitives/chain-rococo" }
 bp-runtime = { path = "../../primitives/runtime" }
+bp-wococo = { path = "../../primitives/chain-wococo" }
 
 bridge-runtime-common = { path = "../../bin/runtime-common" }
 # Substrate Dependencies
@@ -26,4 +28,5 @@ sp-core = { git = "https://github.com/paritytech/substrate", branch = "master" }
 sp-runtime = { git = "https://github.com/paritytech/substrate", branch = "master" }
 
 [dev-dependencies]
+bp-polkadot-core = { path = "../../primitives/polkadot-core" }
 sp-finality-grandpa = { git = "https://github.com/paritytech/substrate", branch = "master" }
