+++ relays/client-bridge-hub-wococo/Cargo.toml
@@ -14,6 +14,8 @@ relay-substrate-client = { path = "../client-substrate" }
 
 bp-bridge-hub-wococo = { path = "../../primitives/chain-bridge-hub-wococo" }
 bp-messages = { path = "../../primitives/messages" }
+bp-parachains = { path = "../../primitives/parachains" }
+bp-runtime = { path = "../../primitives/runtime" }
 relay-bridge-hub-rococo-client = { path = "../client-bridge-hub-rococo" }
 
 # Substrate Dependencies
